Add dense MLP (enable GPU) + speed up with lower float precision

+++ SET-MLP-MPI/utils/load_data.py
@@ -88,8 +88,8 @@ def load_cifar10_data(n_training_samples, n_testing_samples):
     x_train = (x_train - x_train_mean) / x_train_std
     x_test = (x_test - x_train_mean) / x_train_std
 
-    x_train = x_train.reshape(-1, 32 * 32 * 3).astype('float64')
-    x_test = x_test.reshape(-1, 32 * 32 * 3).astype('float64')
+    x_train = x_train.reshape(-1, 32 * 32 * 3).astype('float32')
+    x_test = x_test.reshape(-1, 32 * 32 * 3).astype('float32')
 
     return x_train, y_train, x_test, y_test
